requirements.txt
# ==================== INTERNLINK PYTHON DEPENDENCIES ====================
# Install all dependencies with: pip install -r requirements.txt

# Flask - Web framework for Python
Flask==3.0.0

# Flask-CORS - Handle Cross-Origin Resource Sharing
flask-cors==4.0.0

# MySQL Connector - Database connection
mysql-connector-python==8.2.0

# Werkzeug - Security utilities (included with Flask, but listed for clarity)
Werkzeug==3.0.1

# ==================== OPTIONAL DEPENDENCIES ====================
# Uncomment if you want to use these features

# Python-dotenv - For environment variables (recommended for production)
# python-dotenv==1.0.0

# PyJWT - For JSON Web Tokens (if implementing token-based auth)
# PyJWT==2.8.0

# ==================== INSTALLATION INSTRUCTIONS ====================
# 
# 1. Make sure Python 3.7+ is installed:
#    python --version
#
# 2. Install all dependencies:
#    pip install -r requirements.txt
#
# 3. Verify installation:
#    pip list
#
# ==================== NOTES ====================
# - Use virtual environment (recommended):
#   python -m venv venv
#   source venv/bin/activate  # On Windows: venv\Scripts\activate
#   pip install -r requirements.txt
#
# - Update packages:
#   pip install --upgrade -r requirements.txt
#
# - Freeze current environment:
#   pip freeze > requirements.txt
# Navigate to your project folder
cd path/to/internlink

# Install all packages from requirements.txt
pip install -r requirements.txt